+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod api;
+pub mod api_client;
 pub mod commands;
 pub mod config;
 pub mod error;
